Add if_exists option to drops, fix bind_tables bug
The first to prevent needing to do a try catch if you're not sure if the
table exists, since this way is way cheaper. The bug with
DatabaseInterface.bind_tables was that if you called it before you were
connected to the database, it would throw an exception, since it tried
to get some information about the dialect. Now we just make something
reasonable up if the database isn't connected yet.